requirements.txt
streamlit
groq
pandas
python-dotenv
PyPDF2
audio-recorder-streamlit
gTTS
